requirements.txt
langchain==0.3.13
langchain-google-genai==2.0.5
langgraph==0.2.59
pypdf2==3.0.1
python-dotenv==1.0.1
gradio==5.9.1
pydantic==2.10.6
